refactor: user item component
